Optimize WCOW/LCOW snapshotters to not make a scratch vhdx every snapshot
For WCOW/LCOW currently we create or copy a vhdx for every single snapshot
so there ends up being a sandbox.vhdx in every directory seemingly unnecessarily.
With the default scratch size of 20GB the size on disk is about 17MB so there's a 17MB
overhead per layer plus the time to copy the file with every snapshot. Only the final sandbox.vhd is
actually used so this would be a nice little optimization.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>